+++ frontend/common/base/src/commonMain/kotlin/net/matsudamper/money/frontend/common/base/notification/NotificationAccessState.kt
@@ -1,0 +1,7 @@
+package net.matsudamper.money.frontend.common.base.notification
+
+public sealed interface NotificationAccessState {
+    public data object Granted : NotificationAccessState
+
+    public data object NotGranted : NotificationAccessState
+}
