+++ frontend/common/base/src/commonMain/kotlin/net/matsudamper/money/frontend/common/base/notification/NotificationUsageAccessGateway.kt
@@ -1,0 +1,9 @@
+package net.matsudamper.money.frontend.common.base.notification
+
+import kotlinx.coroutines.flow.Flow
+
+public interface NotificationUsageAccessGateway {
+    public fun accessStateFlow(): Flow<NotificationAccessState>
+
+    public fun openAccessSettings()
+}
